Reuse UserDto mapping
