+++ app/src/main/kotlin/com/olderwold/sliide/data/UsersResponseDTO.kt
@@ -2,31 +2,14 @@ package com.olderwold.sliide.data
 
 import com.google.gson.annotations.SerializedName
 
-data class UsersResponseDTO(
+internal data class UsersResponseDTO(
     @SerializedName("code")
     val code: Int? = null,
     @SerializedName("data")
-    val data: List<Data?> = emptyList(),
+    val data: List<UserDto?> = emptyList(),
     @SerializedName("meta")
     val meta: Meta? = null
 ) {
-    data class Data(
-        @SerializedName("created_at")
-        val createdAt: String? = null,
-        @SerializedName("email")
-        val email: String? = null,
-        @SerializedName("gender")
-        val gender: String? = null,
-        @SerializedName("id")
-        val id: String? = null,
-        @SerializedName("name")
-        val name: String? = null,
-        @SerializedName("status")
-        val status: String? = null,
-        @SerializedName("updated_at")
-        val updatedAt: String? = null
-    )
-
     data class Meta(
         @SerializedName("pagination")
         val pagination: Pagination? = null
